optimize: channel mezon meet on mobile
