+++ apps/mobile/src/app/screens/home/homedrawer/components/ChannelVoice/HeaderRoomView/index.tsx
@@ -8,9 +8,9 @@ import { DeviceEventEmitter, NativeModules, Platform, Text, TouchableOpacity, Vi
 import MezonIconCDN from '../../../../../../componentUI/MezonIconCDN';
 import { IconCDN } from '../../../../../../constants/icon_cdn';
 import AudioOutputTooltip from '../../AudioOutputTooltip';
-import ReactionSoundEffect from '../../EmojiPicker/StickerSelector/ReactionSoundEffect';
 import { ContainerMessageActionModal } from '../../MessageItemBS/ContainerMessageActionModal';
 import { style } from '../styles';
+import SendVoiceSound from './SendVoiceSound';
 import SwitchCamera from './SwitchCamera';
 const { AudioSessionModule } = NativeModules;
 
@@ -23,13 +23,11 @@ export type AudioOutput = {
 type headerProps = {
 	isShow: boolean;
 	channelId: string;
-	clanId: string;
 	onPressMinimizeRoom: () => void;
 	isGroupCall?: boolean;
-	sendSoundReaction: (soundId: string) => void;
 };
 
-const HeaderRoomView = memo(({ channelId, clanId, onPressMinimizeRoom, isGroupCall = false, isShow, sendSoundReaction }: headerProps) => {
+const HeaderRoomView = memo(({ channelId, onPressMinimizeRoom, isGroupCall = false, isShow }: headerProps) => {
 	const { themeValue } = useTheme();
 	const styles = style(themeValue);
 	const [currentAudioOutput, setCurrentAudioOutput] = useState<string>('earpiece');
@@ -130,18 +128,6 @@ const HeaderRoomView = memo(({ channelId, clanId, onPressMinimizeRoom, isGroupCa
 		return channel?.channel_label || '';
 	}, [channelId]);
 
-	const handleSoundSelect = useCallback(
-		async (soundId: string) => {
-			try {
-				sendSoundReaction(soundId);
-				DeviceEventEmitter.emit(ActionEmitEvent.ON_TRIGGER_BOTTOM_SHEET, { isDismiss: true });
-			} catch (error) {
-				console.error('Error sending sound effect:', error);
-			}
-		},
-		[sendSoundReaction]
-	);
-
 	const handleOpenEmojiPicker = () => {
 		const data = {
 			snapPoints: ['45%', '75%'],
@@ -154,16 +140,6 @@ const HeaderRoomView = memo(({ channelId, clanId, onPressMinimizeRoom, isGroupCa
 					channelId={channelId}
 				/>
 			),
-			containerStyle: { zIndex: 1001 },
-			backdropStyle: { zIndex: 1000, backgroundColor: 'rgba(0, 0, 0, 0.3)' }
-		};
-		DeviceEventEmitter.emit(ActionEmitEvent.ON_TRIGGER_BOTTOM_SHEET, { isDismiss: false, data });
-	};
-
-	const handleOpenSoundEffect = () => {
-		const data = {
-			snapPoints: ['45%', '75%'],
-			children: <ReactionSoundEffect onSelected={handleSoundSelect} />,
 			containerStyle: { zIndex: 1001 },
 			backdropStyle: { zIndex: 1000, backgroundColor: 'rgba(0, 0, 0, 0.3)' }
 		};
@@ -189,11 +165,7 @@ const HeaderRoomView = memo(({ channelId, clanId, onPressMinimizeRoom, isGroupCa
 						<MezonIconCDN icon={IconCDN.reactionIcon} height={size.s_20} width={size.s_20} color={themeValue.white} />
 					</TouchableOpacity>
 				)}
-				{!isGroupCall && (
-					<TouchableOpacity onPress={handleOpenSoundEffect} style={[styles.buttonCircle]}>
-						<MezonIconCDN icon={IconCDN.activityIcon} height={size.s_20} width={size.s_20} color={themeValue.white} />
-					</TouchableOpacity>
-				)}
+				{!isGroupCall && <SendVoiceSound channelId={channelId} />}
 				<SwitchCamera />
 				<AudioOutputTooltip
 					onSelectOutput={switchAudioOutput}
