+++ apps/mobile/src/app/screens/home/homedrawer/components/ChannelVoice/index.tsx
@@ -1,13 +1,11 @@
 import { AudioSession, LiveKitRoom, TrackReference, useConnectionState } from '@livekit/react-native';
 import { size, useTheme } from '@mezon/mobile-ui';
-import { getStore, selectChannelById2, selectIsPiPMode, selectVoiceInfo, useAppDispatch, useAppSelector, voiceActions } from '@mezon/store-mobile';
+import { selectIsPiPMode, selectVoiceInfo, useAppDispatch, useAppSelector, voiceActions } from '@mezon/store-mobile';
 import React, { memo, useEffect, useMemo, useRef, useState } from 'react';
 import { AppState, NativeModules, Platform, StatusBar, StyleSheet, View } from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
 import { PERMISSIONS, request } from 'react-native-permissions';
 import { useSelector } from 'react-redux';
-import { ReactionChannelInfo } from '../../../../../../../../../libs/components/src/lib/components/VoiceChannel/MyVideoConference/Reaction/types';
-import { useSendReaction } from '../../../../../../../../../libs/components/src/lib/components/VoiceChannel/MyVideoConference/Reaction/useSendReaction';
 import StatusBarHeight from '../../../../../components/StatusBarHeight/StatusBarHeight';
 import { useSoundReactions } from '../../../../../hooks/useSoundReactions';
 import { CallReactionHandler } from './CallReactionHandler';
@@ -120,12 +118,7 @@ function ChannelVoice({
 	const channelId = useMemo(() => {
 		return voiceInfo?.channelId;
 	}, [voiceInfo]);
-	const currentChannel = useMemo(() => {
-		const store = getStore();
-		const channel = selectChannelById2(store.getState(), channelId);
-		return channel;
-	}, [channelId]);
-	const { sendSoundReaction } = useSendReaction({ currentChannel: currentChannel as ReactionChannelInfo });
+
 	const { handleSoundReaction, activeSoundReactions } = useSoundReactions();
 	const clanId = useMemo(() => {
 		return voiceInfo?.clanId;
@@ -255,11 +248,9 @@ function ChannelVoice({
 				<LiveKitRoom serverUrl={serverUrl} token={token} connect={true}>
 					<HeaderRoomView
 						channelId={channelId}
-						clanId={clanId}
 						onPressMinimizeRoom={onPressMinimizeRoom}
 						isGroupCall={isGroupCall}
 						isShow={isAnimationComplete && !focusedScreenShare && !isPiPMode}
-						sendSoundReaction={sendSoundReaction}
 					/>
 					<ConnectionMonitor />
 					{!isGroupCall && !isPiPMode && isAnimationComplete && (
